nij-platform/backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-environ==0.11.2
psycopg2-binary==2.9.9
Pillow==10.4.0
pdfplumber==0.11.4
pytesseract==0.3.13
celery==5.4.0
redis==5.0.8
gunicorn==22.0.0
whitenoise==6.7.0
drf-spectacular==0.27.2
requests==2.32.3
